Made home layout and added react router to our index.tsx
