+++ package.json
@@ -36,7 +36,7 @@
     "prettier": "2.6.1",
     "react": "^18.0.2",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.11.2",
+    "react-router-dom": "^6.23.1",
     "typescript": "^5.3.3",
     "vite": "^5.0.0",
     "vitest": "^1.6.0"
